Notify TransformationListener before releasing TransformationJob
I have seen the release() call hang in certain cases when processing failed as a result of which the listener
callback never invoked which caused our internal logic to get stuck as well.
I think it's better to first invoke the callback so that LiTr users can at least somehow propagate the error to end users or log the associated exception.